Add unsized impls for StableHashEq and move trait to new file.
